💦 Refactor: 메이트 팝업 창 코드 common으로 분리
